refactor: drive-proof-verifier reorg
