+++ packages/rs-drive-proof-verifier/src/lib.rs
@@ -6,11 +6,12 @@ pub mod uniffi_bindings;
 mod error;
 /// Implementation of proof verification
 pub mod proof;
+pub mod verify;
 
 pub use error::Error;
 #[cfg(feature = "mocks")]
-pub use proof::from_proof::MockQuorumInfoProvider;
-pub use proof::from_proof::{FromProof, QuorumInfoProvider};
+pub use proof::MockQuorumInfoProvider;
+pub use proof::{FromProof, QuorumInfoProvider};
 
 #[cfg(feature = "uniffi")]
 uniffi::include_scaffolding!("dash_drive_v0");
