--- packages/rs-drive-proof-verifier/src/proof/mod.rs
@@ -1,2 +1,0 @@
-pub mod from_proof;
-mod verify;
