+++ packages/rs-drive-proof-verifier/src/verify.rs
@@ -8,8 +8,26 @@ use tenderdash_abci::{
 
 use crate::Error;
 
-use super::from_proof::QuorumInfoProvider;
+use crate::proof::QuorumInfoProvider;
 
+/// Verify cryptographic proof generated by Tenderdash
+///
+/// This function verifies the cryptographic proof generated by Tenderdash.
+/// It verifies the signature of the proof using the public key of the quorum.
+///
+/// ## Parameters
+///
+/// - `proof`: The cryptographic proof generated by Tenderdash
+/// - `mtd`: The metadata of the response from Dash Platform
+/// - `root_hash`: The root hash of the response from Dash Platform
+/// - `provider`: The provider of quorum information
+///
+/// ## Returns
+///
+/// Returns:
+/// * `Ok(())` when the proof is valid
+/// * `Err(Error)` when the proof is invalid
+///
 pub(crate) fn verify_tenderdash_proof(
     proof: &Proof,
     mtd: &ResponseMetadata,
