+++ packages/rs-sdk/src/platform/fetch_many.rs
@@ -7,7 +7,7 @@
 
 use dapi_grpc::platform::v0::GetDocumentsResponse;
 use dpp::document::Document;
-use drive_proof_verifier::proof::from_proof::FromProof;
+use drive_proof_verifier::FromProof;
 use rs_dapi_client::{transport::TransportRequest, DapiRequest, RequestSettings};
 
 use crate::mock::MockResponse;
